arch: elevate to 64-bit Gold Standard with intptr_t
Replaced legacy '(long)' pointer punning with standard 'intptr_t'
throughout the core shell logic. This ensures pointer stability
on 64-bit Linux and prevents truncation during address arithmetic.

- Updated Rcheat, relstak, and round macros.
- Standardized stack offset calculations in name.c and service.c.
- Fixed int-to-pointer cast warning in macro.c.

+++ mode.h
@@ -23,12 +23,14 @@ typedef short BOOL;
 #define	NIL	((char*)0)
 
 
+#include <stdint.h>
+
 /* the following nonsense is required
  * because casts turn an Lvalue
  * into an Rvalue so two cheats
  * are necessary, one for each context.
  */
-#define Rcheat(a)	((long)(a))
+#define Rcheat(a)	((intptr_t)(a))
 
 
 /* address puns for storage allocation */
